Practica Diseño espacios y chorradas

+++ DS-24-08-PD/src/e1/dto/Billete.java
@@ -3,12 +3,12 @@ package e1.dto;
 import java.util.Objects;
 
 public class Billete{
-    private final double precio;
+    private final float precio;
     private final String origen;
     private final String destino;
     private final String fecha;
 
-    public Billete(double precio,String origen,String destino,String fecha) throws IllegalArgumentException {
+    public Billete(float precio,String origen,String destino,String fecha) throws IllegalArgumentException {
         if (origen == null || destino == null || fecha == null) throw new IllegalArgumentException("parametro null");
         if (precio <= 0.0) throw new IllegalArgumentException("El precio no puede ser negativo o 0");
         this.precio=precio;
@@ -17,7 +17,7 @@ public class Billete{
         this.fecha=fecha;
     }
 
-    public double getPrecio() {
+    public float getPrecio() {
         return precio;
     }
 
@@ -38,7 +38,8 @@ public class Billete{
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Billete billete = (Billete) o;
-        return Double.compare(billete.precio, precio) == 0 && origen.equals(billete.origen) && destino.equals(billete.destino) && fecha.equals(billete.fecha);
+        return Float.compare(billete.precio, precio) == 0 && origen.equals(billete.origen) &&
+                destino.equals(billete.destino) && fecha.equals(billete.fecha);
     }
 
     @Override
